ai-service/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.18.2
SQLAlchemy==1.4.39
psycopg2-binary>=2.9.9
python-dotenv==0.20.0
grpcio>=1.58.0,<2.0.0
protobuf==5.29.0
requests>=2.28.1
pydantic>=2.7.4,<3.0.0
groq==0.23.1
ollama==0.3.3
langchain-core==0.3.56
langchain-ollama>=0.2.0,<1.0.0